Fixes for issues reported in #40 and #58
Added client_id for non embedded signature requests to allow the use of
client_id for white-labelling

Updated logic to select correct API endpoint. The call was based on
client_id not being null. This meant that all requests with client_id
populated would create an embedded request. Updated
send_signature_request_embedded parameters to include identifier
to determine the request type. Altered selection process in
_send_signature_request to reflect this.

Included fix for issue #58. The latest merge meant that iuf the option
signing_options was empty this would cause a malformed JSON exception.
The json.dumps call was causing us to send null to the endpoint which
returned a 400 error. Updated the call to replace a null payload with
an empty JSON payload.

Fix for reported issue #40
Fix for reported issue #58