Fix for Issue #26
Passing a list in the from_fields_to_document would result in an error
as the data was in the in correct format. This change converts the
passed data to JSON format before the request is sent.

A utility method has also been added to perform this conversion as it
happens several times throughout the SDK